requirements.txt
streamlit
pandas
numpy
Pillow
torch
torchvision
scikit-learn
matplotlib
